# docs/openapi/registry.openapi.json
{
  "components": {
    "parameters": {
      "AgentIdPath": {
        "description": "Registry agent id.",
        "in": "path",
        "name": "id",
        "required": true,
        "schema": { "minLength": 1, "type": "string" }
      },
...
    },
    "responses": {
      "BadRequest": {
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
        },
        "description": "Request validation failed."
      },
      "Forbidden": {
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
        },
        "description": "The caller cannot access the requested resource."
      },
      "NotFound": {
        "content": {
          "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
        },
        "description": "The requested resource was not found."
      },
...
      "Unauthorized": {
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/AuthErrorResponse" }
          }
        },
        "description": "Control-plane authentication failed."
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "registry token or JWT",
        "description": "Control-plane bearer authentication configured by registrationToken or auth middleware.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
...
  "info": {
    "description": "Machine-readable contract for A2A Mesh registry discovery, control-plane, metrics, and event-stream endpoints.",
    "license": { "identifier": "Apache-2.0", "name": "Apache-2.0" },
    "title": "A2A Mesh Registry API",
    "version": "1.0.0"
  },
...
  "openapi": "3.1.0",
  "paths": {
    "/admin/agents/{id}": {
      "delete": {
        "operationId": "adminDeleteRegistryAgent",
        "responses": {
          "204": { "description": "Agent deleted." },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        },
        "security": [{ "bearerAuth": [] }],
        "summary": "Delete a registered agent through the admin route alias.",
        "tags": ["Admin", "Agents"]
      },
      "parameters": [{ "$ref": "#/components/parameters/AgentIdPath" }]
    },
    "/admin/agents/{id}/heartbeat": {
      "parameters": [{ "$ref": "#/components/parameters/AgentIdPath" }],
      "post": {
        "operationId": "adminHeartbeatRegistryAgent",
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/RegisteredAgent" } }
            },
            "description": "Updated registered agent."
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        },
        "security": [{ "bearerAuth": [] }],
        "summary": "Mark a registered agent healthy through the admin route alias.",
        "tags": ["Admin", "Agents"]
      }
    },
    "/admin/agents/export": {
      "get": {
        "operationId": "exportRegistryAgents",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegistryExportDocument" }
              }
            },
            "description": "Registry export document."
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        },
        "security": [{ "bearerAuth": [] }],
        "summary": "Export registered agents as a registry export document.",
        "tags": ["Admin"]
      }
    },
    "/admin/agents/import": {
      "post": {
        "operationId": "importRegistryAgents",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegistryExportDocument" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegistryImportResult" }
              }
            },
            "description": "Registry import result."
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        },
        "security": [{ "bearerAuth": [] }],
        "summary": "Import registered agents from a registry export document.",
        "tags": ["Admin"]
      }
    },
    "/admin/agents/register": {
      "post": {
        "operationId": "adminRegisterRegistryAgent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterAgentRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/RegisteredAgent" } }
            },
            "description": "The registered agent record."
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        },
        "security": [{ "bearerAuth": [] }],
        "summary": "Register or update an agent through the admin route alias.",
        "tags": ["Admin", "Agents"]
      }
    },
    "/agents": {
      "get": {
        "description": "When public=true is supplied this endpoint returns public agents without control-plane authentication. Otherwise it requires the registry control-plane bearer token or JWT middleware.",
        "operationId": "listRegistryAgents",
        "parameters": [
          { "$ref": "#/components/parameters/PublicQuery" },
          { "$ref": "#/components/parameters/LimitQuery" },
          { "$ref": "#/components/parameters/CursorQuery" }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": { "$ref": "#/components/schemas/RegisteredAgent" },
                  "type": "array"
                }
              }
            },
            "description": "Registered agents visible to the caller.",
            "headers": {
              "X-A2A-Registry-Page-Count": {
                "description": "Number of registered agents returned in this response body.",
                "schema": { "minimum": 0, "type": "integer" }
              },
              "X-A2A-Registry-Page-Next-Cursor": {
                "description": "Cursor to request the next page. Omitted when there is no next page.",
                "schema": { "type": "string" }
              },
              "X-A2A-Registry-Page-Total": {
                "description": "Total number of matching registered agents before pagination.",
                "schema": { "minimum": 0, "type": "integer" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        },
        "security": [{ "bearerAuth": [] }, {}],
        "summary": "List registered agents.",
        "tags": ["Agents"]
      }
    },
    "/agents/{id}": {
      "delete": {
        "operationId": "deleteRegistryAgent",
        "responses": {
          "204": { "description": "Agent deleted." },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        },
        "security": [{ "bearerAuth": [] }],
        "summary": "Delete a registered agent by id.",
        "tags": ["Agents"]
      },
      "get": {
        "description": "Public agents can be fetched without authentication. Private agents require control-plane authentication and tenant access.",
        "operationId": "getRegistryAgent",
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/RegisteredAgent" } }
            },
            "description": "Registered agent."
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        },
        "security": [{ "bearerAuth": [] }, {}],
        "summary": "Fetch a registered agent by id.",
        "tags": ["Agents"]
      },
      "parameters": [{ "$ref": "#/components/parameters/AgentIdPath" }]
    },
    "/agents/{id}/heartbeat": {
      "parameters": [{ "$ref": "#/components/parameters/AgentIdPath" }],
      "post": {
        "operationId": "heartbeatRegistryAgent",
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/RegisteredAgent" } }
            },
            "description": "Updated registered agent."
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        },
        "security": [{ "bearerAuth": [] }],
        "summary": "Mark a registered agent healthy and refresh heartbeat metadata.",
        "tags": ["Agents"]
      }
    },
    "/agents/register": {
      "post": {
        "operationId": "registerRegistryAgent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterAgentRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/RegisteredAgent" } }
            },
            "description": "The registered agent record."
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        },
        "security": [{ "bearerAuth": [] }],
        "summary": "Register or update an agent in the registry.",
        "tags": ["Agents"]
      }
    },
...
    "/events": {
      "get": {
        "operationId": "streamRegistryEvents",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "description": "Server-sent events whose data payload matches RegistryEvent.",
                  "type": "string"
                },
                "x-a2a-event-payload-schema": { "$ref": "#/components/schemas/RegistryEvent" }
              }
            },
            "description": "SSE stream containing registry_update events."
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        },
        "security": [{ "bearerAuth": [] }],
        "summary": "Stream authenticated registry update events.",
        "tags": ["Events"]
      }
    },
...
  },
  "servers": [{ "description": "Local registry server", "url": "http://localhost:3000" }],
...
}